Latest generated openwire
git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/activemq/trunk@384828 13f79535-47bb-0310-9956-ffa450edef68

+++ openwire-cpp/src/command/Response.hpp
@@ -49,7 +49,7 @@ namespace apache
 class Response : public BaseCommand
 {
 private:
-    short correlationId ;
+    int correlationId ;
 
 public:
     const static int TYPE = 30;
@@ -60,8 +60,8 @@ public:
 
     virtual int getCommandType() ;
 
-    virtual short getCorrelationId() ;
-    virtual void setCorrelationId(short correlationId) ;
+    virtual int getCorrelationId() ;
+    virtual void setCorrelationId(int correlationId) ;
 
 
 } ;
